Re-write Get_Giso_u.py
Re-write final calculation iterations so that it sums over velocity for each row of output.txt, then moves on to the next row of output.txt. This means that there will be 2000 outputs for Giso and each will be the result of a Reimann sum over the velocity list for a specific row of output.txt. I am uncertain if this is the way that we wat Giso_u calculated or if we only want one velocity to a row and the sum of those rows.