File to calculate the sigmas and D_theta values.
This file calculates the sigmas and D_theta values used in calculating Giso_u and Gani across all slabs and k values of the reionization front model. The output is stored in txt files named D_theta.txt and sigmas.txt. We will use these to visualize the relationship between D_theta and the sigmas as well as between the sigmas and Giso_u.