Change get_sigmas function.
Update the get_sigmas function so that it prevents unwanted behavior for i*D_theta/kv under i*1e-3. We determined that in the limit of small D_theta/kv, sigma_11 goes to sqrt((3*pi**3)/8) so we forced the value of sigma_11 to be equal to this if D_theta/kv<1e-3.